pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "chee_speech"
version = "0.2.0"
description = "Utils to create an argentine speech corpus with extensive metadata."
authors = [
    {name = "Matías Di Bernardo",email = "matias.di.bernardo@hotmail.com"},
    {name = "Ignacio Correa",email = "n.correa.213@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "pesto-pitch",       
  "pymcd",             
  "pyroomacoustics",
  "pyyaml",
  "tqdm",
  "pandas",
  "openai-whisper",
  "jiwer"
]
